requirements.txt
streamlit
dotenv
google-generativeai
google.adk